# config/openapi.yml
openapi: 3.0.3
info:
  title: Cutshort Social Media / TODO App
  description: API declaration of social media app
  contact:
    email: vk.rahul318@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
  - url: /api/v1
tags:
  - name: todo
    description: TODO APIs
  - name: post
    description: POST APIs
  - name: comment
    description: COMMENT APIs

paths:
  /todo:
    get:
      description: Returns todo
      tags:
        - todo
      operationId: getTodo
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          required: true
          in: query
          description: The userId of a caller
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: string
      responses:
        200:
          description: Success

  /newTodo:
    post:
      description: Posts a new todo to the database
      tags:
        - todo
      operationId: postTodo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - text
                - userId
              properties:
                text:
                  type: string
                userId:
                  type: string
      responses:
        201:
          description: Success
        500:
          description: Error

  /updateTodoText:
    patch:
      description: Updates the text of the todo
      tags:
        - todo
      operationId: updateTodoText
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - text
                - todoId
              properties:
                text:
                  type: string
                todoId:
                  type: string
      responses:
        201:
          description: Success
        500:
          description: Error

  /updateTodoStatus:
    patch:
      description: Marks it as complete
      tags:
        - todo
      operationId: updateTodoStatus
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - todoId
                - completed
              properties:
                todoId:
                  type: string
                completed:
                  type: boolean
      responses:
        201:
          description: Success
        500:
          description: Error

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer